+++ rtp_llm/cpp/cache/KvCacheInfo.h
@@ -6,11 +6,11 @@
 namespace rtp_llm {
 
 struct KVCacheInfo {
-    size_t               available_kv_cache = 0;
-    size_t               total_kv_cache     = 0;
-    size_t               block_size         = 0;
-    std::vector<int64_t> cached_keys;
-    int64_t              version = -1;
+    size_t              available_kv_cache = 0;
+    size_t              total_kv_cache     = 0;
+    size_t              block_size         = 0;
+    std::vector<size_t> cached_keys;
+    int64_t             version = -1;
 };
 
 }  // namespace rtp_llm
